Add option to skip additional certificate validations
This option is sometimes also called `ssl-no-verify` and allows clients to accept certificates that have incorrect values.

We need to connect to a lot of endpoints that we cannot manage ourselves and we get all kinds of errors with the certificates. Fixing those errors is difficult as we do not always have the required access ourselves, yet the endpoints are all 100% trusted. So this options allows us to connect dispite of those errors.