fix: clean up TCP state processing
The runtime component for the connection task shadowed
the `id` variable and was only unregistered in one branch.
So I decided to remove it instead.

The `tokio-stream` documentation says:
> Receiving None indicates that stream iteration is terminated.
So I adapted the places where the read and write tasks tried to
continue after reading `None` from their channels. On my system
this would lead to a lot of CPU activity from inifinite loops.

`spawn_finished_monitor_task` would update the `finished_flag` -
but no one was actually interested in that variable anymore.

The TCP transport tried to call `on_session_closed` when the
connection was closed. After encountering a deadlock for
`session_state` a couple of times, I decided that since session
and connection state should be kept seperate, I might try to
remove that part.